Retrienving serial_read data on ZMQClientComponent::init()

+++ applications/plugins/ZeroMQCommunication/SerialDriver.h
@@ -163,7 +163,7 @@ public:
     void cleanup();
     virtual void draw(const core::visual::VisualParams*) override;
     virtual void draw();
-    int askDevice();
+    float askDevice();
 
     void setForceFeedback(ForceFeedback* ff);
